+++ backend/app/src/main/java/com/bls/patronage/StudyBox.java
@@ -14,6 +14,7 @@ import com.bls.patronage.resources.DeckResource;
 import com.bls.patronage.resources.DecksResource;
 import com.bls.patronage.resources.FlashcardResource;
 import com.bls.patronage.resources.FlashcardsResource;
+import com.bls.patronage.resources.ResetPasswordResource;
 import com.bls.patronage.resources.ResultsResource;
 import com.bls.patronage.resources.TipResource;
 import com.bls.patronage.resources.TipsResource;
@@ -73,7 +74,9 @@ public class StudyBox extends Application<StudyBoxConfiguration> {
         environment.jersey().register(new FlashcardResource(jdbi.onDemand(FlashcardDAO.class)));
         environment.jersey().register(new FlashcardsResource(jdbi.onDemand(FlashcardDAO.class)));
         environment.jersey().register(new UserResource(jdbi.onDemand(UserDAO.class)));
-        environment.jersey().register(new UsersResource(jdbi.onDemand(UserDAO.class), jdbi.onDemand(TokenDAO.class)));
+        environment.jersey().register(new UsersResource(jdbi.onDemand(UserDAO.class)));
+        environment.jersey().register(new ResetPasswordResource(jdbi.onDemand(UserDAO.class),
+                jdbi.onDemand(TokenDAO.class)));
         environment.jersey().register(new TipResource(jdbi.onDemand(TipDAO.class)));
         environment.jersey().register(new TipsResource(jdbi.onDemand(TipDAO.class)));
         environment.jersey().register(new ResultsResource(jdbi.onDemand(FlashcardDAO.class),
